requirements.txt
django
python-decouple
pillow
django-mathfilters
django-crispy-forms
crispy-bootstrap5
django-email-verification